Add promise rejection after timeout
If an error occurs after an application has been launched to get the window size, the corresponding promise may never be resolved. Therefore, a timeout is added to the promise for all window sizes. In case of a timeout, the promise gets rejected. This can be used to trigger another attempt to get the window sizes.